+++ AMglTF/AMglTF Shared/Assets.xcassets/kizunaai/bottoms.textureset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "textures" : [
+    {
+      "idiom" : "universal",
+      "filename" : "Universal.mipmapset"
+    }
+  ]
+}
