+++ AMglTF/AMglTF Shared/Assets.xcassets/kizunaai/eye.textureset/Universal.mipmapset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "levels" : [
+    {
+      "filename" : "eye.png",
+      "mipmap-level" : "base"
+    }
+  ]
+}
