+++ AMglTF/AMglTF Shared/Assets.xcassets/kizunaai/eye2.textureset/Universal.mipmapset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "levels" : [
+    {
+      "filename" : "eye2.png",
+      "mipmap-level" : "base"
+    }
+  ]
+}
