+++ AMglTF/AMglTF Shared/Assets.xcassets/kizunaai/face.textureset/Universal.mipmapset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "levels" : [
+    {
+      "filename" : "face.png",
+      "mipmap-level" : "base"
+    }
+  ]
+}
