+++ AMglTF/AMglTF Shared/Assets.xcassets/kizunaai/hair2.textureset/Universal.mipmapset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  },
+  "levels" : [
+    {
+      "filename" : "hair2.png",
+      "mipmap-level" : "base"
+    }
+  ]
+}
