+++ AMglTF/AMglTF Shared/Utility/Model.swift
@@ -1,0 +1,9 @@
+//
+//  Model.swift
+//  AMglTF
+//
+//  Created by hyunjun529 on 14/09/2018.
+//  Copyright © 2018 hyunjun529. All rights reserved.
+//
+
+import Foundation
